move away from collection api
now app has session model with `sections`
