+++ app/models/session.js
@@ -1,0 +1,7 @@
+import { Model, hasMany } from 'sofa';
+
+export default Model.extend({
+
+  sections: hasMany('section', { inverse: null, relationship: 'session-sections' })
+
+});
